requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.5
django-autocomplete-light==2.3.1
django-crispy-forms==1.6.1
idna==2.6
lxml==4.0.0
mysqlclient==1.3.12
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
